// src/src/models/profile.ts
import { Address } from './address';
import { Education } from './education';
import { Experience } from './experience';
import { Skills } from './skills';

export interface Profile {
    address?: Address;
    company?: any;
    department?: string;
    education: Education[];
    email: string;
    endingRatePerHour: number;
    experience: Experience[];
    firstName: string;
    fullName?: string;
    id: string;
    isTextSubscriber: boolean;
    lastName: string;
    phoneNumber: string;
    photos: any[];
    projectDescription?: string;
    ratePerHour: number;
    role: string;
    skills: Skills[];
    additionalsSkills: Skills[];
    startingRatePerHour: number;
    summary?: string;
    title?: string;
    typeOfWork: string;
    isProfilePublic: boolean;
    isAvailableToHire: boolean;
    resume: File;
    titles: any[];
    jobFunction: string;
    experienceLevel: string;
    culture: string;
    personality: string;
    language: string;
    localisation: string;
    sizeOfWork: string;
    industry: string;
    selectLanguageEng: string;
    selectLanguageFr: string;
    unWantedIndustry: string;
    legallyWork: boolean;
    relocate: boolean;
    questionaire?: any;
    notified?: string;
    processWithCompany: boolean;
    needSponsor: boolean;
    salaryId: string;
    incorporatedHelp: boolean;
    incorporated: boolean;
    codeLangue: string;
    typeSearch?: string;
    searchOpportunity?: string;
    duration?: any[] | string;
    percentage?: {
        personalInformation: number;
        professionalInformation: number;
        myPreferences: number;
        paymentInformation?: number;
    };
}